+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 * `combine_into_workbook()`: Titles and footnotes are now styled again. (18.04.2026)
+* `compute()`: If a variable was all NA, a type miss match happend, leading to the function beeing aborted. This is fixed now. (24.04.2026)
 
 ### Additionally